+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.3.40] - 2022-10-18
+### Changed
+ARGUS-31387
+- Improved error handling in KafkaToHazelcastHandler and HazelcastTransactionalConsumer.
+- Changed default value for `keepThreadAliveOnException` to `true`.
+- Changed `bulk.failed.count` metric to only report if a configurable error threshold is exceeded
+(default threshold is 3, change with `permittedConsecutiveErrors` option).
+- Changed `queue.offer.error.count` metric to `queue.offer.timeout.count`.
+- Added additional metrics `bulk.accepted.count` and `bulk.rejected.count`.
+
 ## [0.3.39] - 2022-10-06
 ### Changed
 ARGUS-31292